+++ Acl/ThreadAclInterface.php
@@ -37,4 +37,17 @@ interface ThreadAclInterface
      * @return void
      */
     function setDefaultAcl(ThreadInterface $thread);
+
+    /**
+     * Installs the Default 'fallback' Acl entries for generic access.
+     *
+     * @return void
+     */
+    function installFallbackAcl();
+
+    /**
+     * Removes default Acl entries
+     * @return void
+     */
+    function uninstallFallbackAcl();
 }
