+++ Document/CommentManager.php
@@ -13,10 +13,8 @@ use Doctrine\ODM\MongoDB\DocumentManager;
 use FOS\CommentBundle\Model\CommentManager as BaseCommentManager;
 use FOS\CommentBundle\Model\ThreadInterface;
 use FOS\CommentBundle\Model\CommentInterface;
-use FOS\CommentBundle\Model\Tree;
-use InvalidArgumentException;
 use DateTime;
-use Exception;
+use InvalidArgumentException;
 
 class CommentManager extends BaseCommentManager
 {
@@ -35,35 +33,6 @@ class CommentManager extends BaseCommentManager
         $this->dm         = $dm;
         $this->repository = $dm->getRepository($class);
         $this->class      = $dm->getClassMetadata($class)->name;
-    }
-
-    /*
-     * Returns all thread comments in a nested array
-     * Will typically be used when it comes to display the comments.
-     *
-     * @param  string $identifier
-     * @return array(
-     *     0 => array(
-     *         'comment' => CommentInterface,
-     *         'children' => array(
-     *             0 => array (
-     *                 'comment' => CommentInterface,
-     *                 'children' => array(...)
-     *             ),
-     *             1 => array (
-     *                 'comment' => CommentInterface,
-     *                 'children' => array(...)
-     *             )
-     *         )
-     *     ),
-     *     1 => array(
-     *         ...
-     *     )
-     */
-    public function findCommentTreeByThread(ThreadInterface $thread, $depth = null)
-    {
-        $comments = $this->findCommentsByThread($thread, $depth);
-        return $this->organiseComments($comments);
     }
 
     /**
@@ -128,36 +97,22 @@ class CommentManager extends BaseCommentManager
         if (null !== $comment->getId()) {
             throw new InvalidArgumentException('Can not add already saved comment');
         }
+
         if (null === $comment->getThread()) {
             throw new InvalidArgumentException('The comment must have a thread');
         }
+
         if (null !== $parent) {
             $comment->setAncestors($this->createAncestors($parent));
         }
+
         $thread = $comment->getThread();
         $thread->setNumComments($thread->getNumComments() + 1);
         $thread->setLastCommentAt(new DateTime());
+
+        $this->dm->persist($thread);
         $this->dm->persist($comment);
         $this->dm->flush();
-    }
-
-    /**
-     * Creates the ancestor array for a given parent
-     * Gets the parent ancestors, and adds the parent id.
-     *
-     * @param CommentInterface $parent
-     * @return array
-     * @throw InvalidArgumentException if the parent has no ID
-     */
-    private function createAncestors(CommentInterface $parent)
-    {
-        if (!$parent->getId()) {
-            throw new InvalidArgumentException('The comment parent must have an ID.');
-        }
-        $ancestors = $parent->getAncestors();
-        $ancestors[] = $parent->getId();
-
-        return $ancestors;
     }
 
     /**
