+++ Model/CommentManager.php
@@ -10,6 +10,8 @@
 
 namespace FOS\CommentBundle\Model;
 
+use InvalidArgumentException;
+
 /**
  * Abstract Comment Manager implementation which can be used as base class for your
  * concrete manager.
@@ -31,21 +33,70 @@ abstract class CommentManager implements CommentManagerInterface
         return $comment;
     }
 
+    /*
+     * Returns all thread comments in a nested array
+     * Will typically be used when it comes to display the comments.
+     *
+     * @param  string $identifier
+     * @return array(
+     *     0 => array(
+     *         'comment' => CommentInterface,
+     *         'children' => array(
+     *             0 => array (
+     *                 'comment' => CommentInterface,
+     *                 'children' => array(...)
+     *             ),
+     *             1 => array (
+     *                 'comment' => CommentInterface,
+     *                 'children' => array(...)
+     *             )
+     *         )
+     *     ),
+     *     1 => array(
+     *         ...
+     *     )
+     */
+    public function findCommentTreeByThread(ThreadInterface $thread, $depth = null)
+    {
+        $comments = $this->findCommentsByThread($thread, $depth);
+
+        return $this->organiseComments($comments);
+    }
+
+    /**
+     * Creates the ancestor array for a given parent
+     * Gets the parent ancestors, and adds the parent id.
+     *
+     * @param CommentInterface $parent
+     * @return array
+     * @throw InvalidArgumentException if the parent has no ID
+     */
+    protected function createAncestors(CommentInterface $parent)
+    {
+        if (!$parent->getId()) {
+            throw new InvalidArgumentException('The comment parent must have an ID.');
+        }
+        $ancestors = $parent->getAncestors();
+        $ancestors[] = $parent->getId();
+
+        return $ancestors;
+    }
+
     /**
      * Organises a flat array of comments into a Tree structure. For
      * organising comment branches of a Tree, certain parents which
      * have not been fetched should be passed in as an array to
      * $ignoreParents.
      *
-     * @param string $comments An array of comments to organise
-     * @param string $ignoreParents An array of parents to ignore
+     * @param array $comments An array of comments to organise
+     * @param array $ignoreParents An array of parents to ignore
      * @return array A tree of comments
      */
     protected function organiseComments($comments, $ignoreParents = null)
     {
         $tree = new Tree();
 
-        foreach($comments as $index => $comment) {
+        foreach($comments as $comment) {
             $path = $tree;
 
             $ancestors = $comment->getAncestors();
